+++ extensions/easydict/src/easydict.tsx
@@ -2,7 +2,7 @@
  * @author: tisfeng
  * @createTime: 2022-06-23 14:19
  * @lastEditor: tisfeng
- * @lastEditTime: 2023-01-08 17:26
+ * @lastEditTime: 2023-03-01 11:19
  * @fileName: easydict.tsx
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
@@ -43,7 +43,7 @@ export default function (props: LaunchProps<{ arguments: EasydictArguments }>) {
   }
 
   const { queryText } = props.arguments;
-  const trimQueryText = queryText ? trimTextLength(queryText) : undefined;
+  const trimQueryText = queryText ? trimTextLength(queryText) : props.fallbackText;
 
   const [isLoadingState, setLoadingState] = useState<boolean>(false);
   const [isShowingDetail, setIsShowingDetail] = useState<boolean>(false);
@@ -100,6 +100,7 @@ export default function (props: LaunchProps<{ arguments: EasydictArguments }>) {
     if (inputText === undefined) {
       setup();
     }
+
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [inputText]);
 
@@ -107,21 +108,23 @@ export default function (props: LaunchProps<{ arguments: EasydictArguments }>) {
    * Do something setup when the extension is activated. Only run once.
    */
   function setup() {
-    console.log(`setup when extension is activated.`);
+    // console.log(`setup when extension is activated.`);
 
     if (trimQueryText?.length) {
-      console.log(`---> arguments queryText: ${trimQueryText}`);
+      console.warn(`---> arguments queryText: ${trimQueryText}`);
     }
 
     const startTime = Date.now();
 
+    const userInputText = trimQueryText;
+
     // If enabled system proxy, we need to wait for the system proxy to be ready.
     if (myPreferences.enableSystemProxy) {
       // If has arguments, use arguments text as input text first.
-      if (trimQueryText?.length) {
+      if (userInputText?.length) {
         configAxiosProxy().then(() => {
           console.log(`after config proxy`);
-          updateInputTextAndQueryText(trimQueryText, false);
+          updateInputTextAndQueryText(userInputText as string, false);
         });
       } else if (myPreferences.enableAutomaticQuerySelectedText) {
         Promise.all([getSelectedText(), configAxiosProxy()])
@@ -141,8 +144,8 @@ export default function (props: LaunchProps<{ arguments: EasydictArguments }>) {
         configAxiosProxy();
       }
     } else {
-      if (trimQueryText?.length) {
-        updateInputTextAndQueryText(trimQueryText, false);
+      if (userInputText?.length) {
+        updateInputTextAndQueryText(userInputText, false);
       } else if (myPreferences.enableAutomaticQuerySelectedText) {
         querySelecedtText().then(() => {
           console.log(`after query selected text`);
@@ -230,6 +233,7 @@ export default function (props: LaunchProps<{ arguments: EasydictArguments }>) {
   }
 
   function onInputChange(text: string) {
+    // console.warn(`onInputChange: ${text}`);
     updateInputTextAndQueryText(text, true);
   }