+++ extensions/github/src/search-repositories.tsx
@@ -10,17 +10,26 @@ import { ExtendedRepositoryFieldsFragment } from "./generated/graphql";
 import { useHistory } from "./helpers/repository";
 import { getGitHubClient } from "./helpers/withGithubClient";
 
+type SearchRepositoriesPrefs = {
+  includeForks: boolean;
+  includeArchived: boolean;
+  displayOwnerName: boolean;
+};
+
 function SearchRepositories() {
   const { github } = getGitHubClient();
 
-  const preferences = getPreferenceValues<{ includeForks: boolean }>();
+  const preferences = getPreferenceValues<SearchRepositoriesPrefs>();
 
   const [searchText, setSearchText] = useState("");
   const [searchFilter, setSearchFilter] = useState<string | null>(null);
 
   const { data: history, visitRepository } = useHistory(searchText, searchFilter);
   const query = useMemo(
-    () => `${searchFilter} ${searchText} fork:${preferences.includeForks}`,
+    () =>
+      `${searchFilter} ${searchText} fork:${preferences.includeForks} ${
+        preferences.includeArchived ? "" : "archived:false"
+      }`,
     [searchText, searchFilter]
   );