+++ extensions/height/src/create_task.tsx
@@ -1,0 +1,6 @@
+import CreateTask from "./components/CreateTask";
+import { CreateTaskFormValues } from "./types/task";
+
+export default function Command({ draftValues }: { draftValues?: CreateTaskFormValues }) {
+  return <CreateTask draftValues={draftValues} />;
+}