+++ extensions/height/src/search_tasks.tsx
@@ -1,0 +1,5 @@
+import SearchTasks from "./components/SearchTasks";
+
+export default function Command() {
+  return <SearchTasks />;
+}