+++ extensions/gitlab/CHANGELOG.md
@@ -1,5 +1,13 @@
 # GitLab Changelog
 
+## [Add Copy Clone Url] - 2023-02-24
+
+- Added command to copy clone url of a project to clipboard
+
+## [Allow Failure Jobs] - 2023-02-24
+
+- Added indicators for jobs that are allowed to fail
+
 ## [Project Create New Issue] - 2023-02-03
 
 - Added command to cerate new issue on a project