+++ extensions/internet-radio/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Internet Radio Changelog
 
+## [Additional Station Metadata] - 2023-02-22
+
+- Add station logos to detail view
+- Add support for several additional station detail fields
+    - Images
+    - Alternate Sites
+    - Alternate Streams
+    - Outdated Sites
+    - Outdated Streams
+    - Social Profiles
+    - Call Signs
+    - Contacts
+    - Policies
+- Add additional genres
+- Fix filtering by genre not updating station count in search field
+- Fix error on play/pause while Menubar Radio command is disabled
+
 ## [Search Stations Database, Import/Export, and More] - 2023-02-08
 
 - Add "Search Stations" command
@@ -13,6 +30,6 @@
 
 ## ["Toggle Play/Pause" Command] - 2023-02-05
 
-- Added "Toggle Play/Pause" Command
+- Add "Toggle Play/Pause" Command
 
 ## [Initial Version] - 2023-01-31