+++ extensions/lega-serie-a/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Lega Serie A Changelog
 
+## [Improvements] - 2023-02-21
+
+- Added Club details
+- Refine Fixtures & Results (live fixtures status, broadcasters)
+- Added Coppa Italia & International Competitions
+
 ## [Initial Version] - 2023-02-10
 
 - Added Standings