+++ extensions/supabase-docs/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Supabase Docs Changelog
 
+## [Update links]- 2023-02-15
+
+Update links to match the new supabase docs site.
+
 ## [Initial Version] - 2022-04-30