+++ extensions/tinypng/CHANGELOG.md
@@ -1,5 +1,9 @@
 # TinyPNG Changelog
 
+## [New Features] - 2023-02-11
+
+Added options to overwrite images.
+
 ## [Improvement] - 2022-09-12
 
 Changed API calls from serial to parallel.